Create version 1.0

+++ WarcraftdleAPI/WarcraftdleAPI.Application/Services/AffiliationService.cs
@@ -9,50 +9,57 @@ namespace WarcraftdleAPI.Application.Services;
 
 public class AffiliationService(WarcraftdleDbContext dbContext)
 {
-	public async Task<IEnumerable<Affiliation>> GetAsync()
-	{
-		var affiliations = await dbContext.Affiliation.ToListAsync();
+    public async Task<IEnumerable<Affiliation>> GetAsync()
+    {
+        var affiliations = await dbContext.Affiliation.ToListAsync();
 
-		return affiliations;
-	}
+        return affiliations;
+    }
 
-	public async Task<Affiliation> GetByIdAsync(int id)
-	{
-		var affiliation = await dbContext.Affiliation.FirstOrDefaultAsync(x => x.Id == id)
-			?? throw new ApiException($"Affiliation with id {id} could not be found", HttpStatusCode.NotFound);
+    public async Task<Affiliation> GetByIdAsync(int id)
+    {
+        var affiliation = await dbContext.Affiliation.FirstOrDefaultAsync(x => x.Id == id)
+            ?? throw new ApiException($"Affiliation with id {id} could not be found", HttpStatusCode.NotFound);
 
-		return affiliation;
-	}
+        return affiliation;
+    }
 
-	public async Task<int> AddAsync(AddAffiliationRequest request)
-	{
-		var affiliation = new Affiliation { Name = request.Name };
+    public async Task<int> AddAsync(AddAffiliationRequest request)
+    {
+        var affiliation = new Affiliation { Name = request.Name };
 
-		await dbContext.Affiliation.AddAsync(affiliation);
-		await dbContext.SaveChangesAsync();
+        await dbContext.Affiliation.AddAsync(affiliation);
+        await dbContext.SaveChangesAsync();
 
-		return affiliation.Id;
-	}
+        return affiliation.Id;
+    }
 
-	public async Task AddMultipleAsync(AddMultipleAffiliationRequest request)
-	{
-		var affiliations = new List<Affiliation>();
+    public async Task AddMultipleAsync(AddMultipleAffiliationRequest request)
+    {
+        var affiliations = new List<Affiliation>();
 
-		foreach (var name in request.AffiliationNames)
-		{
-			affiliations.Add(new Affiliation { Name = name });
-		}
+        foreach (var name in request.AffiliationNames)
+        {
+            affiliations.Add(new Affiliation { Name = name });
+        }
 
-		await dbContext.Affiliation.AddRangeAsync(affiliations);
-		await dbContext.SaveChangesAsync();
-	}
+        await dbContext.Affiliation.AddRangeAsync(affiliations);
+        await dbContext.SaveChangesAsync();
+    }
 
-	public async Task DeleteAsync(int id)
-	{
-		var affiliation = await dbContext.Affiliation.FirstOrDefaultAsync(x => x.Id == id)
-			?? throw new ApiException($"Affiliation with id {id} could not be found", HttpStatusCode.NotFound);
+    public async Task DeleteAsync(int id)
+    {
+        var affiliation = await dbContext.Affiliation
+            .Include(a => a.WowCharacters)
+            .FirstOrDefaultAsync(x => x.Id == id)
+            ?? throw new ApiException($"Affiliation with id {id} could not be found", HttpStatusCode.NotFound);
 
-		dbContext.Affiliation.Remove(affiliation);
-		await dbContext.SaveChangesAsync();
-	}
+        if (affiliation.WowCharacters.Any())
+        {
+            throw new ApiException($"Affiliation with id {id} cannot be deleted because it is used by one or more characters", HttpStatusCode.BadRequest);
+        }
+
+        dbContext.Affiliation.Remove(affiliation);
+        await dbContext.SaveChangesAsync();
+    }
 }
