+++ WarcraftdleAPI/WarcraftdleAPI.Application/Services/ZoneService.cs
@@ -9,50 +9,57 @@ namespace WarcraftdleAPI.Application.Services;
 
 public class ZoneService(WarcraftdleDbContext dbContext)
 {
-	public async Task<IEnumerable<Zone>> GetAsync()
-	{
-		var zones = await dbContext.Zone.ToListAsync();
+    public async Task<IEnumerable<Zone>> GetAsync()
+    {
+        var zones = await dbContext.Zone.ToListAsync();
 
-		return zones;
-	}
+        return zones;
+    }
 
-	public async Task<Zone> GetByIdAsync(int id)
-	{
-		var zone = await dbContext.Zone.FirstOrDefaultAsync(x => x.Id == id)
-			?? throw new ApiException($"Zone with id {id} could not be found", HttpStatusCode.NotFound);
+    public async Task<Zone> GetByIdAsync(int id)
+    {
+        var zone = await dbContext.Zone.FirstOrDefaultAsync(x => x.Id == id)
+            ?? throw new ApiException($"Zone with id {id} could not be found", HttpStatusCode.NotFound);
 
-		return zone;
-	}
+        return zone;
+    }
 
-	public async Task<int> AddAsync(AddZoneRequest request)
-	{
-		var zone = new Zone { Name = request.Name };
+    public async Task<int> AddAsync(AddZoneRequest request)
+    {
+        var zone = new Zone { Name = request.Name };
 
-		await dbContext.Zone.AddAsync(zone);
-		await dbContext.SaveChangesAsync();
+        await dbContext.Zone.AddAsync(zone);
+        await dbContext.SaveChangesAsync();
 
-		return zone.Id;
-	}
+        return zone.Id;
+    }
 
-	public async Task AddMultipleAsync(AddMultipleZoneRequest request)
-	{
-		var zones = new List<Zone>();
+    public async Task AddMultipleAsync(AddMultipleZoneRequest request)
+    {
+        var zones = new List<Zone>();
 
-		foreach (var name in request.ZoneNames)
-		{
-			zones.Add(new Zone { Name = name });
-		}
+        foreach (var name in request.ZoneNames)
+        {
+            zones.Add(new Zone { Name = name });
+        }
 
-		await dbContext.Zone.AddRangeAsync(zones);
-		await dbContext.SaveChangesAsync();
-	}
+        await dbContext.Zone.AddRangeAsync(zones);
+        await dbContext.SaveChangesAsync();
+    }
 
-	public async Task DeleteAsync(int id)
-	{
-		var zone = await dbContext.Zone.FirstOrDefaultAsync(x => x.Id == id)
-			?? throw new ApiException($"Zone with id {id} could not be found", HttpStatusCode.NotFound);
+    public async Task DeleteAsync(int id)
+    {
+        var zone = await dbContext.Zone
+            .Include(z => z.WowCharacters)
+            .FirstOrDefaultAsync(x => x.Id == id)
+            ?? throw new ApiException($"Zone with id {id} could not be found", HttpStatusCode.NotFound);
 
-		dbContext.Zone.Remove(zone);
-		await dbContext.SaveChangesAsync();
-	}
+        if (zone.WowCharacters.Any())
+        {
+            throw new ApiException($"Zone with id {id} cannot be deleted because it is used by one or more characters", HttpStatusCode.BadRequest);
+        }
+
+        dbContext.Zone.Remove(zone);
+        await dbContext.SaveChangesAsync();
+    }
 }
